+++ react/B2BAdminLayout.tsx
@@ -1,0 +1,3 @@
+import { B2BAdminLayout } from 'vtex.b2b-organizations'
+
+export default B2BAdminLayout