+++ react/CostCenterDetails.tsx
@@ -1,0 +1,3 @@
+import { CostCenterDetails } from 'vtex.b2b-organizations'
+
+export default CostCenterDetails