+++ react/Customers.tsx
@@ -1,0 +1,3 @@
+import { index } from 'vtex.admin-customers'
+
+export default index