+++ react/OrganizationRequestDetails.tsx
@@ -1,0 +1,3 @@
+import { OrganizationRequestDetails } from 'vtex.b2b-organizations'
+
+export default OrganizationRequestDetails