+++ react/roleEdit.tsx
@@ -1,0 +1,3 @@
+import { roleEdit } from 'vtex.storefront-permissions-ui'
+
+export default roleEdit