+++ react/roleNew.tsx
@@ -1,0 +1,3 @@
+import { roleNew } from 'vtex.storefront-permissions-ui'
+
+export default roleNew