+++ react/rolesList.tsx
@@ -1,0 +1,3 @@
+import { rolesList } from 'vtex.storefront-permissions-ui'
+
+export default rolesList