+++ react/userList.tsx
@@ -1,0 +1,3 @@
+import { userList } from 'vtex.storefront-permissions-ui'
+
+export default userList